Colorpicker margaritov overwork and fixes which were urgent needed
More or less over a year a lot roms are using margaritov's and nillson's colorpicker
Due to the progress of roms this colorpicker needed a lot overwork now. Some fixes, some
improvements etc. etc. Should now be bugfree + additions we added and add now

Before I do a bigger list screenshot here....should explain all:

https://dl.dropbox.com/u/106208842/Screenshot_2013-02-26-02-39-02.png
https://dl.dropbox.com/u/106208842/Screenshot_2013-02-26-02-39-11.png

Important hints for kangers. Due that aosp depreciated hex color values on this stage
I decided to remove also in colorpicker this support in general to be aosp consistent.
This means that for defaultValues only int colors are accepted. Two ways to do this.

1. Set defaultValue in xml code. Like android:defaultValue="0xff33be5e" or link it there
   to a integer color resource (before there was only the possibility to set eg "#33be5e" which
   is depreciated from aosp)

2. Do it with the new integrated method I did about 2 month ago. setNewPreviewColor(int color)
   Of course at this point you can do to get hex resources like it was before, eg
   setNewPreviewColor(getResources().getColor(com.android.internal.R.color.holo_blue_dark)

3. or do both :p

cheers

Change-Id: Ie8f3d624fb2bd7ded4cef525157545d161cc1af8